fix: isolate Parsl energy result helper
